Exceptions move part7
Problem:
Currently exception classes are scattered all over SDK, this will lead to unecessary duplication and confusion

Analysis:
Removed the following classes from resource.py:

AttemptedMutationOfReadOnly
BooleansToReduceHaveSameValue

Corrected coresponding import statements

Tests:
Flake8
